[refactor] move the router functions into a controller file
